requirements.txt
pandas
torch
numpy

timm
tqdm
sklearn
Pillow
transformers
openpyxl
pydicom
nltk
albumentations